+++ CHANGELOG.md
@@ -12,15 +12,17 @@
 
 - building-comparison: AOI is now clipped to the coverage area ([#739])
 - mapping-saturation: add hover info to estimated total data line ([#723])
+- building-comparison: now has link to reference datasets ([#741])
 
 ### Other Changes
 
 - building-comparison: no quality estimation for areas with strong difference to reference data ([#739])
 - test(db): add missing mock for getting coverage from database in tests ([#739])
-- building-comparison: request EUBUCCO database now much faster ([#743])
+- building-comparison: request EUBUCCO database is now much faster ([#743])
 
 [#723]: https://github.com/GIScience/ohsome-quality-api/pull/723
 [#739]: https://github.com/GIScience/ohsome-quality-api/pull/739
+[#741]: https://github.com/GIScience/ohsome-quality-api/pull/741
 [#743]: https://github.com/GIScience/ohsome-quality-api/pull/743